+++ alt-text/src/components/icons/Lightning.tsx
@@ -1,19 +1,19 @@
 export function Lightning() {
   return (
     <svg
-      fill="none"
-      stroke="currentColor"
-      strokeLinecap="round"
+      data-testid="geist-icon"
+      height="16"
       strokeLinejoin="round"
-      strokeWidth="1.5"
-      style={{
-        height: '1rem',
-        width: '1rem',
-      }}
-      viewBox="0 0 24 24"
-      xmlns="http://www.w3.org/2000/svg"
+      style={{ color: 'currentcolor', height: '1rem', width: '1rem' }}
+      viewBox="0 0 16 16"
+      width="16"
     >
-      <polygon points="13 2 3 14 12 14 11 22 21 10 12 10 13 2" />
+      <path
+        clipRule="evenodd"
+        d="M7.157 0L2.333 9.408l-.56 1.092H7a.25.25 0 01.25.25V16h1.593l4.824-9.408.56-1.092H9a.25.25 0 01-.25-.25V0H7.157zM7 9H4.227L7.25 3.106V5.25C7.25 6.216 8.034 7 9 7h2.773L8.75 12.894V10.75A1.75 1.75 0 007 9z"
+        fill="currentColor"
+        fillRule="evenodd"
+      />
     </svg>
   )
 }