+++ pages/src/icons/HomeIcon.tsx
@@ -9,17 +9,19 @@ export const HomeIcon: React.FC<{
     <svg
       aria-label={ariaLabel}
       className={className}
-      fill="none"
+      data-testid="geist-icon"
       height={size}
-      stroke="currentColor"
-      strokeLinecap="round"
       strokeLinejoin="round"
-      strokeWidth="2"
-      viewBox="0 0 24 24"
+      style={{ color: 'currentcolor' }}
+      viewBox="0 0 16 16"
       width={size}
     >
-      <path d="m3 9 9-7 9 7v11a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2z" />
-      <polyline points="9 22 9 12 15 12 15 22" />
+      <path
+        clipRule="evenodd"
+        d="M12.5 6.56L8 2.06l-4.5 4.5v6.94H6V11a2 2 0 114 0v2.5h2.5V6.56zm1.28-.84L8.707.645a1 1 0 00-1.414 0L2.22 5.72.47 7.47-.06 8 1 9.06l.53-.53.47-.47V15h12V8.06l.47.47.53.53L16.06 8l-.53-.53-1.75-1.75zM8.5 11v2.5h-1V11a.5.5 0 111 0z"
+        fill="currentColor"
+        fillRule="evenodd"
+      />
     </svg>
   )
 }