+++ vercel-deployments/src/components/icons/clock.tsx
@@ -3,6 +3,7 @@ import React from 'react'
 export const ClockIcon: React.FC<React.SVGProps<SVGSVGElement>> = (props) => {
   return (
     <svg
+      data-testid="geist-icon"
       height="16"
       strokeLinejoin="round"
       style={{ color: 'currentcolor' }}
@@ -12,7 +13,7 @@ export const ClockIcon: React.FC<React.SVGProps<SVGSVGElement>> = (props) => {
     >
       <path
         clipRule="evenodd"
-        d="M14.5 8C14.5 11.5899 11.5899 14.5 8 14.5C4.41015 14.5 1.5 11.5899 1.5 8C1.5 4.41015 4.41015 1.5 8 1.5C11.5899 1.5 14.5 4.41015 14.5 8ZM16 8C16 12.4183 12.4183 16 8 16C3.58172 16 0 12.4183 0 8C0 3.58172 3.58172 0 8 0C12.4183 0 16 3.58172 16 8ZM8.75 4.75V4H7.25V4.75V7.875C7.25 8.18976 7.39819 8.48615 7.65 8.675L9.55 10.1L10.15 10.55L11.05 9.35L10.45 8.9L8.75 7.625V4.75Z"
+        d="M14.5 8a6.5 6.5 0 11-13 0 6.5 6.5 0 0113 0zM16 8A8 8 0 110 8a8 8 0 0116 0zM8.75 4.75V4h-1.5v3.875a1 1 0 00.4.8l1.9 1.425.6.45.9-1.2-.6-.45-1.7-1.275V4.75z"
         fill="currentColor"
         fillRule="evenodd"
       />