+++ vercel-deployments/src/components/icons/refresh.tsx
@@ -2,19 +2,19 @@ import React from 'react'
 
 export const RefreshIcon: React.FC<React.SVGProps<SVGSVGElement>> = (props) => (
   <svg
-    fill="none"
+    data-testid="geist-icon"
     height="14"
-    stroke="currentColor"
-    strokeLinecap="round"
     strokeLinejoin="round"
-    strokeWidth="2"
-    viewBox="0 0 24 24"
+    style={{ color: 'currentcolor' }}
+    viewBox="0 0 16 16"
     width="14"
-    xmlns="http://www.w3.org/2000/svg"
     {...props}
   >
-    <polyline points="23 4 23 10 17 10" />
-    <polyline points="1 20 1 14 7 14" />
-    <path d="M3.51 9a9 9 0 0 1 14.85-3.36L23 10M1 14l4.64 4.36A9 9 0 0 0 20.49 15" />
+    <path
+      clipRule="evenodd"
+      d="M8 1.25a6.999 6.999 0 00-6.16 3.672l-.357.66 1.32.713.357-.66a5.503 5.503 0 0110.112 1.039h-2.198v1.5h4.175a.75.75 0 00.75-.75V3.25h-1.5v2.395A7.002 7.002 0 008 1.25zm-6.499 9.605v2.395h-1.5V9.075a.75.75 0 01.75-.75h4.175v1.5H2.729a5.503 5.503 0 0010.098 1.065l.36-.658 1.316.72-.361.659a7.002 7.002 0 01-12.64-.755z"
+      fill="currentColor"
+      fillRule="evenodd"
+    />
   </svg>
 )